[NodeTraverserQueue] improve
