+++ composer.json
@@ -30,6 +30,7 @@
             "Rector\\": "src",
             "Rector\\BetterReflection\\": "packages/BetterReflection/src",
             "Rector\\NodeTypeResolver\\": "packages/NodeTypeResolver/src",
+            "Rector\\NodeTraverserQueue\\": "packages/NodeTraverserQueue/src",
             "Rector\\DeprecationExtractor\\": "packages/DeprecationExtractor/src",
             "Rector\\NodeValueResolver\\": "packages/NodeValueResolver/src"
         }
@@ -38,8 +39,7 @@
         "psr-4": {
             "Rector\\Tests\\": "tests",
             "Rector\\NodeTypeResolver\\Tests\\": "packages/NodeTypeResolver/tests",
-            "Rector\\DeprecationExtractor\\Tests\\": "packages/DeprecationExtractor/tests",
-            "Rector\\NodeValueResolver\\Tests\\": "packages/NodeValueResolver/tests"
+            "Rector\\DeprecationExtractor\\Tests\\": "packages/DeprecationExtractor/tests"
         }
     },
     "scripts": {
