+++ src/Testing/Application/FileProcessor.php
@@ -2,13 +2,9 @@
 
 namespace Rector\Testing\Application;
 
-use PhpParser\Lexer;
 use PhpParser\NodeVisitor;
-use Rector\Contract\Parser\ParserInterface;
-use Rector\NodeTraverser\CloningNodeTraverser;
 use Rector\NodeTraverser\MainNodeTraverser;
-use Rector\NodeTraverser\ShutdownNodeTraverser;
-use Rector\NodeTraverser\StandaloneTraverseNodeTraverser;
+use Rector\NodeTraverserQueue\NodeTraverserQueue;
 use Rector\Printer\FormatPerservingPrinter;
 use Rector\Rector\RectorCollector;
 use SplFileInfo;
@@ -16,19 +12,9 @@ use SplFileInfo;
 final class FileProcessor
 {
     /**
-     * @var ParserInterface
-     */
-    private $parser;
-
-    /**
      * @var FormatPerservingPrinter
      */
     private $codeStyledPrinter;
-
-    /**
-     * @var Lexer
-     */
-    private $lexer;
 
     /**
      * @var MainNodeTraverser
@@ -41,38 +27,20 @@ final class FileProcessor
     private $rectorCollector;
 
     /**
-     * @var CloningNodeTraverser
+     * @var NodeTraverserQueue
      */
-    private $cloningNodeTraverser;
-
-    /**
-     * @var ShutdownNodeTraverser
-     */
-    private $shutdownNodeTraverser;
-
-    /**
-     * @var StandaloneTraverseNodeTraverser
-     */
-    private $standaloneTraverseNodeTraverser;
+    private $nodeTraverserQueue;
 
     public function __construct(
-        CloningNodeTraverser $cloningNodeTraverser,
-        ParserInterface $parser,
         FormatPerservingPrinter $codeStyledPrinter,
-        Lexer $lexer,
         MainNodeTraverser $mainNodeTraverser,
         RectorCollector $rectorCollector,
-        ShutdownNodeTraverser $shutdownNodeTraverser,
-        StandaloneTraverseNodeTraverser $standaloneTraverseNodeTraverser
+        NodeTraverserQueue $nodeTraverserQueue
     ) {
-        $this->parser = $parser;
         $this->codeStyledPrinter = $codeStyledPrinter;
-        $this->lexer = $lexer;
         $this->mainNodeTraverser = $mainNodeTraverser;
         $this->rectorCollector = $rectorCollector;
-        $this->cloningNodeTraverser = $cloningNodeTraverser;
-        $this->shutdownNodeTraverser = $shutdownNodeTraverser;
-        $this->standaloneTraverseNodeTraverser = $standaloneTraverseNodeTraverser;
+        $this->nodeTraverserQueue = $nodeTraverserQueue;
     }
 
     /**
@@ -100,16 +68,9 @@ final class FileProcessor
     /**
      * See https://github.com/nikic/PHP-Parser/issues/344#issuecomment-298162516.
      */
-    public function processFile(SplFileInfo $file): string
+    public function processFile(SplFileInfo $fileInfo): string
     {
-        $oldStmts = $this->parser->parseFile($file->getRealPath());
-        $oldTokens = $this->lexer->getTokens();
-        $newStmts = $this->cloningNodeTraverser->traverse($oldStmts);
-
-        $newStmts = $this->standaloneTraverseNodeTraverser->traverse($newStmts);
-
-        $newStmts = $this->mainNodeTraverser->traverse($newStmts);
-        $newStmts = $this->shutdownNodeTraverser->traverse($newStmts);
+        [$newStmts, $oldStmts, $oldTokens] = $this->nodeTraverserQueue->processFileInfo($fileInfo);
 
         return $this->codeStyledPrinter->printToString($newStmts, $oldStmts, $oldTokens);
     }
