+++ src/config/services.yml
@@ -2,6 +2,7 @@ imports:
   - { resource: '../../packages/BetterReflection/src/config/services.yml' }
   - { resource: '../../packages/DeprecationExtractor/src/config/services.yml' }
   - { resource: '../../packages/NodeTypeResolver/src/config/services.yml' }
+  - { resource: '../../packages/NodeTraverserQueue/src/config/services.yml' }
   - { resource: '../../packages/NodeValueResolver/src/config/services.yml' }
 
 parameters:
